requirements.txt
alumnium
pytest
pytest-playwright
python-dotenv
pytest-html